Short-circuit dependency resolution for already resolved plug-ins
Our simple algorithm was sometimes inserted already-resolved plug-ins
into the array, because we were applying the resolution algorithm to all
plug-ins.

The latter is a necessary evil, because we can't know beforehand which,
if any, plug-in we should start traversing the dependency graph from in
order to cover the entire graph.

Given that we will sometimes be re-walking the a portion of the graph,
add an explicit short-circuit check to avoid adding already-resolved
plug-ins to the list again.

Signed-off-by: Wincent Colaiuta <win@wincent.com>