Remove redundant clean-up from dependency resoluton algorithm
Being in the "unresolved" set is used to detect circular dependencies.
It really means "already seen but not yet resolved", so if you see the
same plug-in again it means that you're looping.

It is, however, not necessary to remove items from the "unresolved" set
upon resolution, because, by definition, the "unresolved" set will never
be checked for items that are in the "resolved" array. That is, in
pseudo-code:

  if (!resolved but seen)
    raise

So, if we've added something to "resolved" the "seen" array won't even
be checked, and if it won't be checked for such items then there's no
need to bother removing them.

Signed-off-by: Wincent Colaiuta <win@wincent.com>